+++ src/components/Home/Authenticated/buttons/index.ts
@@ -1,0 +1,3 @@
+export { NewTripButton } from "./NewTripButton";
+export { JoinTripButton } from "./JoinTripButton";
+export { CopyTripIdButton } from "./CopyTripIdButton";
