+++ src/components/Home/Authenticated/index.ts
@@ -1,6 +1,4 @@
-export { CopyTripIdButton } from "./CopyTripIdButton";
+export { CopyTripIdButton, NewTripButton, JoinTripButton } from "./buttons";
 export { AuthenticatedContainer } from "./AuthenticatedContainer";
 export { TripNav } from "./TripNav";
-export { NewTripButton } from "./NewTripButton";
 export { TripsContainer } from "./TripsContainer";
-export { JoinTripButton } from "./JoinTripButton";
